Update layout
Update the layout and make the code more organised. It's still a mess,
but less so than before.
Added homing keys so it's easier to see where homing is.
Added stl for the build so a prototype can be made.
For some reason OpenSCAD generates a broken STL. Didn't think that was
possible but apparently it is. Will have to investigate.