Add promicro holder and update GB
First: add a more correct model of a pro micro.

Second: create a hodler for a PM that uses friction to keep it in
place. There is a lid on it that helps keep in it place incase it should
wiggle loose for some reason. There is a slight overlap between the lid
and the usb-port as well. If you bend the lid down a little it will have
a slight push on the usb. This helps it sit more snuggly as well as
giving a little pushback for when you insert a cable.

Third: More work on Greta-Britta as a whole. Coming together quite well.
Need to order some parts to make I2C between the halves possible and so
I can measure them out and add to the design.